requirements.txt
# requirements.txt

streamlit>=1.20.0
google-generativeai>=0.4.0 # Keep if using Google models directly
langchain>=0.1.0
langchain-google-genai>=0.0.6 # Keep if using Google models directly
langchain-openai>=0.1.0     # ADDED: For OpenRouter/OpenAI compatible APIs
Pillow>=9.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0
sentence-transformers>=2.2.0
torch>=2.0.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0 # Keep, might still be useful for other env vars
tiktoken>=0.5.0